Rewrite "New Wizard" SWT tests to no longer require the SWTBot runner
Newer Eclipse versions no longer launch a workspace when running a test
outside the UI thread. Because this is a requirement for all of our
SWTBot tests, we can no longer run them locally.

The tests were originally written like that because they open a modal
dialog and would thus lead to a deadlock if the test is also executed
within the UI thread. But thanks to the changes to the `UiContext`, we
have a way to make this work regardless.
