+++ org.eclipse.wb.tests/src/org/eclipse/wb/tests/designer/core/nls/ui/PropertiesCompositeTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2011, 2025 Google, Inc. and others.
+ * Copyright (c) 2011, 2026 Google, Inc. and others.
  *
  * This program and the accompanying materials are made available under the
  * terms of the Eclipse Public License 2.0 which is available at
@@ -16,9 +16,9 @@ import org.eclipse.wb.internal.core.nls.edit.IEditableSource;
 import org.eclipse.wb.internal.core.nls.model.LocaleInfo;
 import org.eclipse.wb.internal.core.nls.ui.NewSourceDialog;
 import org.eclipse.wb.internal.core.nls.ui.PropertiesComposite;
+import org.eclipse.wb.internal.core.utils.execution.ExecutionUtils;
 import org.eclipse.wb.internal.core.utils.reflect.ReflectionUtils;
 import org.eclipse.wb.tests.gef.UiContext;
-import org.eclipse.wb.tests.swtbot.designer.bot.UIUtil;
 import org.eclipse.wb.tests.utils.SWTBotEditableSource;
 
 import static org.eclipse.swtbot.swt.finder.matchers.WidgetOfType.widgetOfType;
@@ -302,8 +302,8 @@ public class PropertiesCompositeTest extends AbstractDialogTest {
 	 */
 	private static SWTBotEditableSource getEditableSource(SWTBot shell) throws Exception {
 		PropertiesComposite composite = shell.getFinder().findControls(widgetOfType(PropertiesComposite.class)).get(0);
-		IEditableSource editableSource = (IEditableSource) UIUtil
-				.syncCall(() -> ReflectionUtils.invokeMethod(composite, "getSelectedSource()"));
+		IEditableSource editableSource = (IEditableSource) UIThreadRunnable.syncExec(
+				() -> ExecutionUtils.runObject(() -> ReflectionUtils.invokeMethod(composite, "getSelectedSource()")));
 		return new SWTBotEditableSource(editableSource);
 	}
 }
