--- org.eclipse.wb.tests/src/org/eclipse/wb/tests/swtbot/designer/WindowBuilder SWTBot Logging.properties
@@ -1,6 +1,0 @@
-handlers= java.util.logging.ConsoleHandler
-
-java.util.logging.ConsoleHandler.level = INFO
-# Based on the example in java.util.logging.SimpleFormatter:
-# This prints 1 line with the log level (4$), the timestamp (1$) in a square bracket and the log message (5$).
-java.util.logging.SimpleFormatter.format = %4$s: [%1$tc] %5$s%n
