--- org.eclipse.wb.tests/src/org/eclipse/wb/tests/swtbot/designer/WindowBuilderTests.java
@@ -1,28 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2024 Patrick Ziegler and others.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * https://www.eclipse.org/legal/epl-2.0.
- *
- * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *    Patrick Ziegler - initial API and implementation
- *******************************************************************************/
-package org.eclipse.wb.tests.swtbot.designer;
-
-import org.eclipse.wb.tests.swtbot.designer.rcp.RcpTests;
-
-import org.junit.platform.suite.api.SelectClasses;
-import org.junit.platform.suite.api.Suite;
-
-/**
- * Executes all SWTBot-based WindowBuilder tests
- */
-@Suite
-@SelectClasses({
-	RcpTests.class,
-})
-public class WindowBuilderTests {
-}
